requirements.txt
# Django and REST
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0

# AWS S3
boto3==1.37.18

# ML & Image Processing
torch==2.6.0
torchvision==0.21.0
numpy==2.2.4
opencv-python-headless==4.11.0.86
Pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2

# Data & Utils
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.10 # PostgreSQL database (used on Render)
matplotlib==3.10.1

# Required by other libraries (dependencies)
typing_extensions==4.12.2
six==1.17.0
urllib3==2.3.0
idna==3.10
certifi==2025.1.31
charset-normalizer==3.4.1
sqlparse==0.5.3
tzdata==2025.1

gunicorn==20.1.0
